Decided to move the networking stuff to its own file
This was largly in an attempt to clean up RouteMatch.java, and to consolidate and fix issues that result from networking quirks.
