+++ app/src/main/java/fnsb/macstransit/RouteMatch/RouteMatch.java
@@ -1,11 +1,5 @@
 package fnsb.macstransit.RouteMatch;
 
-import android.util.Log;
-
-import org.json.JSONObject;
-
-import java.io.BufferedReader;
-import java.io.IOException;
 
 /**
  * Created by Spud on 2019-10-12 for the project: MACS Transit.
@@ -15,7 +9,7 @@ import java.io.IOException;
 public class RouteMatch {
 
 	/**
-	 * The url to pull route data from.
+	 * The feed url to pull route data from.
 	 * <p>
 	 * Example: https://fnsb.routematch.com/feed/
 	 */
@@ -29,7 +23,7 @@ public class RouteMatch {
 	/**
 	 * Constructor for the RouteMatch object.
 	 *
-	 * @param url    The URL to pull data from (IE: https://fnsb.routematch.com/feed/)
+	 * @param url    The feed url to pull data from (IE: https://fnsb.routematch.com/feed/)
 	 * @param routes The array of all the bus routes that are <b><i>able to be</i></b> tracked.
 	 */
 	public RouteMatch(String url, Route[] routes) {
@@ -37,111 +31,6 @@ public class RouteMatch {
 		this.routes = routes;
 	}
 
-	/**
-	 * TODO Documentation
-	 */
-	private static int retryattempts = 0;
-
-	/**
-	 * TODO Update documentation and comments
-	 * Reads the JSON from the provided URL, and formats it into a JSONObject. If the URL times out,
-	 * or responds with an error the method will retry.
-	 *
-	 * @param url The URL to retrieve the JSON data from.
-	 * @return The JSONObject containing the data (or a blank JSON Object if there was an error).
-	 */
-	static JSONObject readJsonFromUrl(String url) {
-
-		Log.d("readJsonFromUrl", url);
-		StringBuilder jsonString = new StringBuilder();
-
-		Thread t = new Thread(() -> {
-
-			try {
-				// Specify the URL connection
-				java.net.URLConnection connection = new java.net.URL(url).openConnection();
-
-				// Add timeouts for the connection (1.5 seconds to connect, 2 seconds to read, 3.5 seconds total)
-				connection.setConnectTimeout(1500);
-				connection.setReadTimeout(2000);
-
-				// Get the input stream from the connection
-				java.io.InputStream inputStream = connection.getInputStream();
-
-				// Create a buffered reader for the input stream
-				BufferedReader bufferedReader = new BufferedReader(new java.io.InputStreamReader(inputStream, java.nio.charset.StandardCharsets.UTF_8));
-
-				// Store the inputted text into a string variable
-				jsonString.append(RouteMatch.readAll(bufferedReader));
-
-				bufferedReader.close();
-				inputStream.close();
-
-			} catch (java.io.FileNotFoundException | java.net.SocketTimeoutException e) {
-				throw new RuntimeException();
-			} catch (IOException uhoh) {
-				uhoh.printStackTrace();
-			}
-
-		});
-		t.setUncaughtExceptionHandler((t1, e) -> {
-			jsonString.setLength(0);
-		});
-		t.setName("Network thread");
-		t.start();
-		try {
-			t.join(4000);
-			if (jsonString.length() == 0) {
-				if (RouteMatch.retryattempts < 3) {
-					// Keep trying!
-					try {
-						Thread.sleep(1000);
-					} catch (InterruptedException ex) {
-						ex.printStackTrace();
-					}
-					RouteMatch.retryattempts++;
-					android.util.Log.w("readJsonFromUrl", String.format("Page didn't respond, going to retry! (%d/3)", RouteMatch.retryattempts));
-					return RouteMatch.readJsonFromUrl(url);
-				} else {
-					RouteMatch.retryattempts = 0;
-					return new JSONObject();
-				}
-			} else {
-				// Create and return a new JSONObject from the jsonString variable
-				RouteMatch.retryattempts = 0;
-				return new JSONObject(jsonString.toString());
-			}
-		} catch (org.json.JSONException | InterruptedException e) {
-			e.printStackTrace();
-			return new JSONObject();
-		}
-	}
-
-	/**
-	 * Parse characters from a reader or buffered reader into a stream.
-	 *
-	 * @param reader The Reader or BufferedReader object.
-	 * @return The final string from the String builder containing what was read by the Reader.
-	 */
-	@Deprecated
-	private static String readAll(java.io.Reader reader) throws IOException {
-
-		// Create a string to store what is read by the reader
-		StringBuilder string = new StringBuilder();
-
-		// Create a variable for character parsing
-		int character;
-
-		// Loop through all the characters until there are no more characters to run through (returns -1)
-		while ((character = reader.read()) != -1) {
-
-			// Append the character to the string
-			string.append((char) character);
-		}
-
-		// Finally return the string
-		return string.toString();
-	}
 
 	/**
 	 * Gets the route data from the url provided in the constructor.
@@ -149,9 +38,9 @@ public class RouteMatch {
 	 * @param routeName The name of the route to get the pertaining data from (IE: Red).
 	 * @return The JSONObject pertaining to that specific route's data.
 	 */
-	public JSONObject getRoute(String routeName) {
+	public org.json.JSONObject getRoute(String routeName) {
 		// Example usage: readJsonFromUrl("https://fnsb.routematch.com/feed/vehicle/byRoutes/Red");
-		return RouteMatch.readJsonFromUrl(this.url + "/vehicle/byRoutes/" + routeName);
+		return fnsb.macstransit.Network.readJsonFromUrl(this.url + "/vehicle/byRoutes/" + routeName);
 	}
 
 	/**
@@ -160,10 +49,10 @@ public class RouteMatch {
 	 * @return A JSONObject array that contains all the individual routes's data.
 	 */
 	@Deprecated
-	public JSONObject[] getAllRoutes() {
+	public org.json.JSONObject[] getAllRoutes() {
 
 		// Create the JSONObject array, and make it the size of the total routes provided in this object
-		JSONObject[] jsonObjects = new JSONObject[this.routes.length];
+		org.json.JSONObject[] jsonObjects = new org.json.JSONObject[this.routes.length];
 
 		// Iterate through each line, and be sure to keep track of the current index
 		for (int index = 0; index < jsonObjects.length; index++) {
